requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-phone-field==1.8.1
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
google-auth==1.20.1
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-decouple==3.3
python-linkedin-v2==0.9.4
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
